+++ tests/framework/db/pgsql/PostgreSQLActiveRecordTest.php
@@ -18,6 +18,23 @@ class PostgreSQLActiveRecordTest extends ActiveRecordTest
 {
     protected $driverName = 'pgsql';
 
+    protected function setUp()
+    {
+        parent::setUp();
+        // load config to use db component
+        $config = self::getParam('databases')['pgsql'];
+        $this->mockApplication([
+            'components' => [
+                'db' => [
+                    'class' => '\yii\db\Connection',
+                    'dsn' => $config['dsn'],
+                    'username' => $config['username'],
+                    'password' => $config['password'],
+                ]
+            ]
+        ]);
+    }
+
     public function testBooleanAttribute()
     {
         /* @var $customerClass \yii\db\ActiveRecordInterface */
@@ -167,6 +184,101 @@ class PostgreSQLActiveRecordTest extends ActiveRecordTest
         $record->type = 'type';
         $record->save(false);
         $this->assertEquals(5, $record->primaryKey);
+    }
+
+    public function testTextArrayAttribute()
+    {
+        $tags = ['programming', 'OOP'];
+        $question = new \yiiunit\data\ar\Question();
+        $question->title = 'How to write a good code?';
+        $question->tags = $tags;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEquals($tags, $question->tags);
+
+        $tags[] = 'creational pattern"s';
+        $question->tags = $tags;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEquals($tags, $question->tags);
+
+        $tags[] = "pattern's";
+        $question->tags = $tags;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEquals($tags, $question->tags);
+
+        $tags[] = "behavioral pattern''s";
+        $question->tags = $tags;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEquals($tags, $question->tags);
+
+        $tags[] = "lazy loading";
+        $question->tags = $tags;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEquals($tags, $question->tags);
+
+        $question->tags = null;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEmpty($question->tags);
+
+        $tags[] = "with ,";
+        $question->tags = $tags;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEquals($tags, $question->tags);
+
+        $tags[] = "with {}";
+        $question->tags = $tags;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEquals($tags, $question->tags);
+    }
+
+    public function testIntegerArrayAttribute()
+    {
+        // ids of related questions
+        $relatedIds = [];
+        $question = new \yiiunit\data\ar\Question();
+        $question->title = 'How to optimize page loading?';
+        $question->tags = ['html', 'css', 'optimization'];
+        $question->related = $relatedIds;
+        $this->assertTrue($question->save(false));
+        $this->assertEquals($relatedIds, $question->related);
+
+        $relatedIds[] = rand(100, 999);
+        $question->related = $relatedIds;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEquals($relatedIds, $question->related);
+    }
+
+    public function testNumericArrayAttribute()
+    {
+        $points = [0.25, 3.8, 4.2];
+        $question = new \yiiunit\data\ar\Question();
+        $question->title = 'How to run unit tests?';
+        $question->points = $points;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEquals($points, $question->points);
+
+        array_pop($points);
+        $question->points = $points;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEquals($points, $question->points);
+
+        $points[] = -2.24;
+        $storedPoints = $question->points;
+        $storedPoints[] = '-2,24';
+        $question->points = $storedPoints;
+        $this->assertTrue($question->save(false));
+        $question->refresh();
+        $this->assertEquals($points, $question->points);
     }
 }